notebooks/requirements.txt
jupyter==1.0.0
ipykernel==6.29.5
ipywidgets==8.1.3
pandas
sqlalchemy
psycopg2-binary
python-dotenv
tqdm
chardet